refactor : store model-domain 분리
