+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/store/service/StoreService.java
@@ -4,18 +4,18 @@ import com.sparta.blackwhitedeliverydriver.store.dto.StoreIdResponseDto;
 import com.sparta.blackwhitedeliverydriver.store.dto.StoreRequestDto;
 import com.sparta.blackwhitedeliverydriver.store.dto.StoreResponseDto;
 import com.sparta.blackwhitedeliverydriver.store.dto.StoreByMasterRequestDto;
-import com.sparta.blackwhitedeliverydriver.category.entity.Category;
-import com.sparta.blackwhitedeliverydriver.store.entity.Store;
-import com.sparta.blackwhitedeliverydriver.store.entity.StoreCategory;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
-import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.model.category.entity.Category;
+import com.sparta.blackwhitedeliverydriver.model.store.entity.Store;
+import com.sparta.blackwhitedeliverydriver.model.store.entity.StoreCategory;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.UserRoleEnum;
 import com.sparta.blackwhitedeliverydriver.category.exception.CategoryExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.store.exception.StoreExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.category.repository.CategoryRepository;
-import com.sparta.blackwhitedeliverydriver.store.repository.StoreCategoryRepository;
-import com.sparta.blackwhitedeliverydriver.store.repository.StoreRepository;
-import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.model.category.repository.CategoryRepository;
+import com.sparta.blackwhitedeliverydriver.model.store.repository.StoreCategoryRepository;
+import com.sparta.blackwhitedeliverydriver.model.store.repository.StoreRepository;
+import com.sparta.blackwhitedeliverydriver.model.user.repository.UserRepository;
 import com.sparta.blackwhitedeliverydriver.common.security.UserDetailsImpl;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
